refactor: refactor session for better SSR
