+++ src/app/api/repos/list/route.ts
@@ -1,0 +1,9 @@
+import { NextResponse } from "next/server";
+
+import repositories from "./src/data/repositories.json";
+
+export const GET = async () => {
+  return NextResponse.json({
+    data: repositories,
+  });
+};
