+++ src/app/api/testimonials/list/route.ts
@@ -1,0 +1,9 @@
+import { NextResponse } from "next/server";
+
+import testimonials from "./src/data/testimonials.json";
+
+export const GET = async () => {
+  return NextResponse.json({
+    data: testimonials,
+  });
+};
