+++ src/app/api/testimonials/list/src/@types/index.ts
@@ -1,0 +1,12 @@
+export type Testimonial = {
+  name: string;
+  role: string;
+  company: string;
+  img: {
+    src: string;
+    alt: string;
+  };
+  relationship: string;
+  content: string;
+  className?: string;
+};
