+++ src/components/organisms/Articles/index.ts
@@ -1,0 +1,3 @@
+import { ArticlesView } from "./src/views/Articles";
+
+export { ArticlesView };
