--- src/components/organisms/Articles/index.tsx
@@ -1,3 +1,0 @@
-import { ArticlesView } from "./src/view/Articles";
-
-export { ArticlesView };
