--- src/components/organisms/Repos/data/repositories.tsx
@@ -1,106 +1,0 @@
-type TechStack = {
-  name: string;
-  color: string;
-  link: string;
-};
-export type Repository = {
-  title: string;
-  cover: string;
-  description_key?: string;
-  project_link?: string;
-  github_link: string;
-  tech_stack: Array<TechStack>;
-};
-export const repositories: Array<Repository> = [
-  {
-    title: "My portfolio",
-    description_key: "my-portfolio",
-    github_link: "https://github.com/GuiMoraesDev/my-portfolio",
-    cover:
-      "https://raw.githubusercontent.com/GuiMoraesDev/my-portfolio/main/public/cover/800p.png",
-    tech_stack: [
-      {
-        name: "Next.js",
-        color: "#000000",
-        link: "https://nextjs.org/",
-      },
-      {
-        name: "Tailwind CSS",
-        color: "#06B6D4",
-        link: "https://tailwindcss.com/",
-      },
-      {
-        name: "Typescript",
-        color: "#3178C6",
-        link: "https://www.typescriptlang.org/",
-      },
-      {
-        name: "Next-intl",
-        color: "#FF0000",
-        link: "https://next-intl-docs.vercel.app/",
-      },
-      {
-        name: "Radix UI",
-        color: "#000000",
-        link: "https://radix-ui.com/",
-      },
-      {
-        name: "OpenAI",
-        color: "#00FF00",
-        link: "https://openai.com/",
-      },
-    ],
-  },
-  {
-    title: "Lnk.bio",
-    description_key: "lnk-bio",
-    github_link: "https://github.com/GuiMoraesDev/lnk.bio",
-    cover: "/projects/lnk.bio.png",
-    tech_stack: [
-      {
-        name: "Next.js",
-        color: "#000000",
-        link: "https://nextjs.org/",
-      },
-      {
-        name: "Tailwind CSS",
-        color: "#06B6D4",
-        link: "https://tailwindcss.com/",
-      },
-      {
-        name: "Typescript",
-        color: "#3178C6",
-        link: "https://www.typescriptlang.org/",
-      },
-    ],
-  },
-  {
-    title: "Prep.pal",
-    description_key: "prep-pal",
-    github_link: "https://github.com/GuiMoraesDev/prep-pal",
-    cover:
-      "https://raw.githubusercontent.com/GuiMoraesDev/prep-pal/main/public/cover.png",
-    tech_stack: [
-      {
-        name: "Next.js",
-        color: "#000000",
-        link: "https://nextjs.org/",
-      },
-      {
-        name: "Tailwind CSS",
-        color: "#06B6D4",
-        link: "https://tailwindcss.com/",
-      },
-      {
-        name: "Typescript",
-        color: "#3178C6",
-        link: "https://www.typescriptlang.org/",
-      },
-      {
-        name: "OpenAI",
-        color: "#00FF00",
-        link: "https://openai.com/",
-      },
-    ],
-  },
-];
