+++ src/components/organisms/Repos/index.ts
@@ -1,0 +1,3 @@
+import { ReposView } from "./src/views/Repos";
+
+export { ReposView };
