+++ src/components/organisms/Testimonials/index.ts
@@ -1,0 +1,3 @@
+import { TestimonialsView } from "./src/views/Testimonials";
+
+export { TestimonialsView };
