+++ src/services/api/api.testimonials.ts
@@ -1,14 +1,14 @@
 import { headers } from "next/headers";
 
-import { type DevDotToArticle } from "@/app/api/articles/list/src/@types";
+import { type Testimonial } from "@/app/api/testimonials/list/src/@types";
 
 const list = async (): Promise<{
-  data: DevDotToArticle[];
+  data: Testimonial[];
 }> => {
   const headersList = await headers();
   const domain = headersList.get("x-current-origin");
 
-  const { href } = new URL(`${domain}/api/articles/list`);
+  const { href } = new URL(`${domain}/api/testimonials/list`);
 
   const response = await fetch(href, {
     method: "GET",
@@ -16,6 +16,6 @@ const list = async (): Promise<{
   return response.json();
 };
 
-export const articles = {
+export const testimonials = {
   list,
 };
