+++ src/services/api/index.ts
@@ -1,5 +1,9 @@
 import { articles } from "./api.articles";
+import { repos } from "./api.repos";
+import { testimonials } from "./api.testimonials";
 
 export const api = {
   articles,
+  testimonials,
+  repos,
 };
